I can't call this a top-down... more like a furious cleanup. Handle lots of buffer overflows. Fix overlength strings in menus. float/double, integer overflow issues. Fix div by zero. Fix null pointer dereferences.

+++ player_points.c
@@ -197,7 +197,7 @@ void check_for_levelup (edict_t *ent)
 			int plateau_level = (int)ceil(log(plateau_points / start_nextlevel->value) / log(nextlevel_mult->value));
 
 			// calculate next level points based
-			points_needed = plateau_points + 5000*(ent->myskills.level-plateau_level);
+			points_needed = plateau_points + 5000 * ((double)ent->myskills.level - (double)plateau_level);
 		}
 
 		ent->myskills.next_level += points_needed;
